Update rs.json
Translation for Serbian language.
If symbols Č, Ć, Š, Đ, Ž are not allowed they can be replaced with C, C, S, Dj/dj, Z
And if Cyrillic letter is supported I can do that too. (simple as rewriting it or using https://www.konvertor.co.rs/)

Some sanding of the translation will be needed.